# docs/swagger.yaml
definitions:
  models.Asset:
    properties:
      configs:
        type: string
      create_at:
        type: integer
      create_by:
        type: string
      dashboard:
        type: string
      directory_id:
        type: integer
      group_id:
        type: integer
      health:
        description: 下面的是健康检查使用，在memsto缓存中保存
        type: integer
      id:
        type: integer
      ident:
        type: string
      ip:
        type: string
      label:
        type: string
      manufacturers:
        type: string
      memo:
        type: string
      name:
        type: string
      optional_metrics:
        description: 巡检检查使用
        items:
          $ref: '#/definitions/models.Metrics'
        type: array
      organization_id:
        type: integer
      params:
        type: string
      plugin:
        type: string
      position:
        type: string
      status:
        description: '0: 未生效, 1: 已生效'
        enum:
        - 0
        - 1
        type: integer
      tags:
        items:
          type: string
        type: array
      type:
        type: string
      update_at:
        type: integer
      update_by:
        type: string
    type: object
...
  models.Metrics:
    properties:
      metrics:
        type: string
      name:
        type: string
    type: object
...
  router.UpdateBody:
    properties:
      ids:
        items:
          type: integer
        type: array
      name:
        type: string
      value: {}
    type: object
  router.assetsModel:
    properties:
      configs:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      ident:
        type: string
      ip:
        type: string
      label:
        type: string
      manufacturers:
        type: string
      memo:
        type: string
      name:
        type: string
      optional_metrics:
        type: string
      organization_id:
        type: integer
      params:
        type: string
      position:
        type: string
      tags:
        type: string
      type:
        type: string
      version:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/n9e/xh/asset/export-xls:
    post:
      consumes:
      - multipart/form-data
      description: EXCEL导出配线架信息
      parameters:
      - description: 导入查询条件
        in: query
        name: query
        type: string
      produces:
      - application/msexcel
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: EXCEL导出配线架信息
      tags:
      - 配线架信息
  /api/n9e/xh/asset/import-xls:
    post:
      consumes:
      - multipart/form-data
      description: EXCEL导入资产
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: EXCEL导入资产
      tags:
      - 资产详情
  /api/n9e/xh/asset/templet:
    post:
      consumes:
      - application/json
      description: 导出资产模板
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Asset'
      security:
      - ApiKeyAuth: []
      summary: 导出资产模板
      tags:
      - 资产详情
...
  /api/n9e/xh/assets/:
    post:
      consumes:
      - application/json
      description: 创建资产-西航
      parameters:
      - description: add assetsModel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.assetsModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 创建资产-西航
      tags:
      - 资产-西航
    put:
      consumes:
      - application/json
      description: 更新资产-西航
      parameters:
      - description: add assetsModel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.assetsModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 更新资产-西航
      tags:
      - 资产-西航
  /api/n9e/xh/assets/batch-del/:
    post:
      consumes:
      - application/json
      description: 批量删除资产-西航
      parameters:
      - description: add assetsModel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.assetsModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 批量删除资产-西航
      tags:
      - 资产-西航
  /api/n9e/xh/assets/batch-update/:
    post:
      consumes:
      - application/json
      description: 批量修改资产属性-西航
      parameters:
      - description: add UpdateBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.UpdateBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 批量修改资产属性-西航
      tags:
      - 资产-西航
  /api/n9e/xh/assets/filter:
    post:
      consumes:
      - application/json
      description: 过滤器-西航
      parameters:
      - description: add query
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 过滤器-西航
      tags:
      - 资产-西航
  /api/n9e/xh/assets/xml/:
    get:
      consumes:
      - application/json
      description: 导出xml
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 导出xml（测试）
      tags:
      - 资产-西航
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"